New test structures (#724)

+++ python/datashaper/tests/test_schema.py
@@ -12,7 +12,7 @@ from pandas.testing import assert_frame_equal
 
 from datashaper import PandasDtypeBackend, Workflow
 
-FIXTURES_PATH = "fixtures/workflow"
+FIXTURES_PATH = "fixtures/workflow/verbs"
 TABLE_STORE_PATH = "fixtures/workflow_inputs"
 SCHEMA_PATH = "workflow.json"
 log = getLogger(__name__)
@@ -25,42 +25,28 @@ thread.start()
 
 
 @cache
-def load_inputs(pandas_dtype_backend: PandasDtypeBackend, engine: str):
+def load_inputs():
     dataframes: dict[str, pd.DataFrame] = {}
+
     for file in os.listdir(TABLE_STORE_PATH):
+        path = Path(TABLE_STORE_PATH) / file
         if file.endswith(".csv"):
-            path = Path(TABLE_STORE_PATH) / file
-            table = pd.read_csv(
-                path,
-                dtype_backend=pandas_dtype_backend,
-                engine=engine,
-            )
+            table = read_csv(path)
             dataframes[file.removesuffix(".csv")] = table
+        if file.endswith(".json") and file != "workflow.json":
+            table = pd.read_json(path)
+            dataframes[file.removesuffix(".json")] = table
     return dataframes
 
 
 def read_csv(path: str) -> pd.DataFrame:
-    table = pd.read_csv(path)
-
-    if "date" in table.columns:
-        table["date"] = pd.to_datetime(table["date"], errors="coerce")
-        table["newColumn"] = pd.to_datetime(table["newColumn"], errors="coerce")
-
-    if "json_object" in table.columns:
-        table["json_object"] = __conversion_function_map["json_object"](
-            table["json_object"]
-        )
-
-    return table
+    return pd.read_csv(path)
 
 
-__conversion_function_map = {
-    "json_object": lambda column: column.apply(
-        lambda _x: json.loads(
-            str(_x).replace("'", '"').replace('"{"', '{"').replace('"}"}', '"}}')
-        )
-    ),
-}
+# pandas won't auto-guess iso dates on import, so we define an explicit iso output for comparison as strings
+# this means that the expected.csv _must_ match the expected iso format to seconds granularity
+def to_csv(df: pd.DataFrame, path: str) -> None:
+    df.to_csv(path, date_format="%Y-%m-%dT%H:%M:%SZ", index=False)
 
 
 def get_verb_test_specs(root: str) -> list[str]:
@@ -74,22 +60,17 @@ def get_verb_test_specs(root: str) -> list[str]:
 @pytest.mark.parametrize("fixture_path", get_verb_test_specs(FIXTURES_PATH))
 @pytest.mark.parametrize(
     "pandas_dtype_backend",
-    [PandasDtypeBackend.NUMPY_NULLABLE, PandasDtypeBackend.PYARROW],
-)
-@pytest.mark.parametrize(
-    "engine",
-    ["python", "pyarrow"],
+    [PandasDtypeBackend.NUMPY_NULLABLE],
 )
 async def test_verbs_schema_input(
     fixture_path: str,
     pandas_dtype_backend: PandasDtypeBackend,
-    engine: str,
 ):
     with (Path(fixture_path) / "workflow.json").open() as schema:
         schema_dict = json.load(schema)
         tables: dict[str, pd.DataFrame] = {}
         for input in schema_dict["input"]:
-            tables[input] = load_inputs(pandas_dtype_backend, engine)[input]
+            tables[input] = load_inputs()[input]
 
         workflow = Workflow(
             schema=schema_dict,
@@ -104,22 +85,27 @@ async def test_verbs_schema_input(
     table_path = str(py_path) if py_path.exists() else fixture_path
 
     for expected in os.listdir(table_path):
-        result_table_path = Path(table_path) / f"result_{expected}"
-        expected_table_path = Path(table_path) / expected
+        if expected != "workflow.json" and expected != "comments.txt":
+            result_table_path = Path(table_path) / f"result_{expected}"
+            expected_table_path = Path(table_path) / expected
+            table_name = expected.split(".")[0]
+            table_name_arg = table_name
 
-        if expected.endswith(".csv"):
+            result = workflow.output(table_name_arg)
+
             try:
-                table_name = expected.split(".")[0]
-                table_name_arg = table_name
-                result = workflow.output(table_name_arg)
-                if isinstance(result, pd.DataFrame):
-                    result.to_csv(result_table_path, index=False)
-                else:
-                    result.obj.to_csv(result_table_path, index=False)
-
-                expected_table = read_csv(expected_table_path)
-                result_table = read_csv(result_table_path)
-
+                # if it's a csv we cycle it through file system to confirm read/write
+                # TODO: this cycle is outdated and should be removed. it masks potential verbs issues
+                if expected.endswith(".csv"):
+                    if isinstance(result, pd.DataFrame):
+                        to_csv(result, result_table_path)
+                    else:
+                        to_csv(result.obj, result_table_path)
+                    expected_table = read_csv(expected_table_path)
+                    result_table = read_csv(result_table_path)
+                elif expected.endswith(".json"):
+                    result_table = result
+                    expected_table = pd.read_json(expected_table_path)
                 assert_frame_equal(
                     expected_table,
                     result_table,
@@ -129,8 +115,9 @@ async def test_verbs_schema_input(
                 )
             except AssertionError:
                 print(  # noqa: T201
-                    f"Error in {fixture_path}@{expected.removesuffix('.csv')};\nExpected:\n{expected_table.head()}\n\nActual:{result_table.head()}",
+                    f"Error in {fixture_path}@{expected};\nExpected:\n{expected_table.head()}\n\nActual:\n{result_table.head()}",
                 )
                 raise
             finally:
-                result_table_path.unlink()
+                if result_table_path.exists():
+                    result_table_path.unlink()
